StudentTerminal/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
Flask-WTF==0.14.2
forex-python==1.5
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.17.4
pkg-resources==0.0.0
pyparsing==2.4.5
python-dateutil==2.8.1
python-dotenv==0.10.3
requests==2.22.0
simplejson==3.17.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
WTForms==2.2.1
